Adding fork of core role
This role is a modified fork (with no history) of the core role.

By its nature the core role is needed for almost all projects, since this role is for internal use only this presents a problem for sharing projects externally (where a key component would be missing).

This core-public role aims to address that by providing a modified version of the core role, with internal information removed, for use publicly.

This role will be maintained seperately but should track the core role wherever possible.