Adding SSl private key copy support
This allows a SSL certificate private key to be copied from the Ansible host to a secure location. By default this is /etc/ssl/private but this, and the name of the copied private key is configurable.

The path and file name of the private key are exposed as variable overrides to the Apache and Nginx roles and requires no changes to these roles to be made.