Refactoring SSH/Sudo file modifications and user modifications
Previously templating was used to update the sudoers file and SSH known_hosts files. This meant a changed event was triggered each time the role was run when in reality, after the first run-through, nothing changed.

To correct this the 'lineinfile' module is used to ensure missing changes are applied, but not applied again if not needed. Since the modifications to these files was so small (a single line per file) this is fine (and preferable).

For users (controller and app) the core role creates two dot files in the user's home directories. Previously these files were made even if they already exist, again causing a change event when no change occured. New tasks using 'stat' are used to first determine if these dotfiles already exist, and is so, to skip over the creation tasks.

This means this role when run a second time will return 'ok' or 'skipped' for all of its tasks. Although these changes seem significant, they don't break backwards compatibility and don't add or change any existing feature (just how they are implemented), therefore these changes are made as a path release.