Removing sudoers support and adding controller user to adm group
This role will no longer modify the sudoers file, you must therefore ensure that any system this role is applied to has the required changes this role did make already applied.

If using a BAS OS image (OVA or Vagrant box) later than version 2.0.0 this will be the case. If using a cloud OS image (e.g. DigitalOcean) you will need to use a snapshot or similar method to use a base, rather than a vanilla OS install. See the relevant bootstrap role for your provider for more information.

As this is a breaking change a new major release has been created, consequently this role now becomes 1.0.0.

Adding the controller user to this group is optional, but enabled by default. It allows the controller to access log files without using sudo for example.